move functions of ExecEngineUtil that used only in test to BlockGenerator
